requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.17.0
redis==5.0.1
sqlalchemy==2.0.25
chromadb==0.4.22
PyPDF2==3.0.1      # For extracting text from PDFs
python-docx==1.1.0 
numpy==2.2.5
faiss-cpu==1.10.0
sentence-transformers==4.1.0
requests==2.32.3
